+++ aeon/classification/tests/test_all_classifiers.py
@@ -8,12 +8,11 @@ import numpy as np
 from sklearn.utils._testing import set_random_state
 
 from aeon.datasets import load_basic_motions, load_unit_test
-from aeon.testing import (
-    BaseFixtureGenerator,
-    QuickTester,
+from aeon.testing.expected_results.expected_classifier_outputs import (
     basic_motions_proba,
     unit_test_proba,
 )
+from aeon.testing.test_all_estimators import BaseFixtureGenerator, QuickTester
 from aeon.utils._testing.estimator_checks import _assert_array_almost_equal
 from aeon.utils._testing.scenarios_classification import ClassifierFitPredict
 from aeon.utils.validation.collection import get_n_cases
