+++ aeon/datasets/tests/test_dataframe_loaders.py
@@ -10,7 +10,7 @@ from aeon.datasets import (
     load_from_tsfile_to_dataframe,
     load_from_ucr_tsv_to_dataframe,
 )
-from aeon.testing import PR_TESTING
+from aeon.testing.test_config import PR_TESTING
 
 
 @pytest.mark.skipif(
