+++ aeon/forecasting/model_evaluation/tests/test_evaluate.py
@@ -31,7 +31,7 @@ from aeon.performance_metrics.forecasting import (
     MeanAbsolutePercentageError,
     MeanAbsoluteScaledError,
 )
-from aeon.testing import PR_TESTING
+from aeon.testing.test_config import PR_TESTING
 from aeon.utils._testing.forecasting import make_forecasting_problem
 from aeon.utils._testing.hierarchical import _make_hierarchical
 from aeon.utils.validation._dependencies import _check_soft_dependencies
