+++ aeon/testing/test_softdeps.py
@@ -15,7 +15,7 @@ from importlib import import_module
 import pytest
 
 from aeon.registry import all_estimators
-from aeon.testing.tests.test_config import EXCLUDE_ESTIMATORS
+from aeon.testing.test_config import EXCLUDE_ESTIMATORS
 from aeon.utils._testing.scenarios_getter import retrieve_scenarios
 from aeon.utils.validation._dependencies import (
     _check_python_version,
