+++ aeon/transformations/tests/test_all_transformers.py
@@ -8,12 +8,11 @@ import pandas as pd
 from sklearn.utils._testing import set_random_state
 
 from aeon.datasets import load_basic_motions, load_unit_test
-from aeon.testing import (
-    BaseFixtureGenerator,
-    QuickTester,
+from aeon.testing.expected_results.expected_transform_outputs import (
     basic_motions_result,
     unit_test_result,
 )
+from aeon.testing.test_all_estimators import BaseFixtureGenerator, QuickTester
 from aeon.utils._testing.estimator_checks import _assert_array_almost_equal
