+++ aeon/utils/_testing/scenarios_forecasting.py
@@ -18,7 +18,7 @@ import pandas as pd
 
 from aeon.base import BaseObject
 from aeon.forecasting.base import BaseForecaster
-from aeon.testing import PR_TESTING
+from aeon.testing.test_config import PR_TESTING
 from aeon.utils._testing.collection import _make_collection_X
 from aeon.utils._testing.hierarchical import _make_hierarchical
 from aeon.utils._testing.scenarios import TestScenario
